% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_events.R
\name{create_events}
\alias{create_events}
\title{Create events for the HEV model}
\usage{
create_events(farms = NULL, days = NULL, named_events = TRUE)
}
\arguments{
\item{farms}{list with farm data.}

\item{days}{the number of days to generate events.}

\item{named_events}{use named events ("exit", "enter", "intTrans",
"extTrans") instead of integers (0, 1, 2, 3). Default is to
use named events.}
}
\value{
a \code{data.frame} with events.
}
\description{
Create events for the HEV model
}
